generation/requirements.txt
langgraph>=0.0.11
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
sentence-transformers==2.7.0
torch==2.2.2
langgraph==0.6.4
langchain==0.3.16
PyMuPDF==1.25.1
langchain-google-genai>=2.0.0,<3
